- refactor various classes into separate files.
